requirements.txt
langchain
sentence-transformers
chromadb
PyMuPDF
streamlit
langchain-community
faiss-cpu